feat: Refactoring hostname
Refactor hostname into variable that can be passed around in
specialArgs. This is probably not the right way to do this, but was the
only way I could get this to work. It allows the hostname variable to be
set once and referenced in multiple modules, that might need to set
options including the specific host the config was generated for. This
also moves the config.networking.hostName variable to the common config
instead of defined in the system specific configs.

Unfortunately, I was unable to figure out how to directly reference
config.networking.hostName. Even when I was able to reference it,
through `config`, it was never actually set to anything (just empty
string). This is also probably a bad way to go about this, but here we
are...